Refactor Game Structure and Implement Dynamic Item Generation
